get rid of duplicated data loading code
