+++ tcga_util.py
@@ -21,61 +21,6 @@ from dask_ml.model_selection import GridSearchCV
 
 import config as cfg
 
-def load_pancancer_data():
-    """Load data to build feature matrices from pancancer repo. """
-
-    base_url = "https://github.com/greenelab/pancancer/raw"
-    commit = "2a0683b68017fb226f4053e63415e4356191734f"
-
-    file = "{}/{}/data/sample_freeze.tsv".format(base_url, commit)
-    sample_freeze_df = pd.read_csv(file, index_col=0, sep='\t')
-
-    file = "{}/{}/data/pancan_mutation_freeze.tsv.gz".format(base_url, commit)
-    mutation_df = pd.read_csv(file, index_col=0, sep='\t')
-
-    file = "{}/{}/data/copy_number_loss_status.tsv.gz".format(base_url, commit)
-    copy_loss_df = pd.read_csv(file, index_col=0, sep='\t')
-
-    file = "{}/{}/data/copy_number_gain_status.tsv.gz".format(base_url, commit)
-    copy_gain_df = pd.read_csv(file, index_col=0, sep='\t')
-
-    file = "{}/{}/data/mutation_burden_freeze.tsv".format(base_url, commit)
-    mut_burden_df = pd.read_csv(file, index_col=0, sep='\t')
-
-    return (
-        sample_freeze_df,
-        mutation_df,
-        copy_loss_df,
-        copy_gain_df,
-        mut_burden_df
-    )
-
-
-def load_top_50():
-    """Load top 50 mutated genes from BioBombe repo.
-
-    These were precomputed for the equivalent experiments in the
-    BioBombe paper, so no need to recompute them.
-    """
-    base_url = "https://github.com/greenelab/BioBombe/raw"
-    commit = "aedc9dfd0503edfc5f25611f5eb112675b99edc9"
-
-    file = "{}/{}/9.tcga-classify/data/top50_mutated_genes.tsv".format(
-            base_url, commit)
-    genes_df = pd.read_csv(file, sep='\t')
-
-    return genes_df
-
-
-def subset_genes_by_mad(train_df, test_df, mad_file, subset_mad_genes):
-    # subset genes by mean absolute deviation
-    mad_genes_df = pd.read_csv(mad_file, sep='\t')
-    mad_genes = mad_genes_df.iloc[0:subset_mad_genes, ].gene_id.astype(str)
-
-    train_df = train_df.reindex(mad_genes, axis='columns')
-    test_df = test_df.reindex(mad_genes, axis='columns')
-    return (train_df, test_df)
-
 
 def build_feature_dictionary(models_dir, load_data=False, store_train_test="both"):
     """
